Pad configuration file to filesystem block boundary.
On recent (4.x?) kernels, the UML virtual block device (ubd) enforces
the filesystem size to be a multiple of a 512-byte block. If the
source file is smaller, it's repeated from the beginning, so a
configuration file containing "foo\n" will instead read as
"foo\nfoo\nfoo\n..." up to 512 bytes. This causes the umlbox init to
execute unnecessary commands (and typically error out) afterwards.

Since init uses 0-terminated strings, padding the file to the block
boundary with 0s makes it behave reasonably.