Simplify the wrapper script file descriptor wrangling.
Instead of whatever the script used to do, consistently use the
following scheme:

The output side of the UML /dev/tty1 ('con1') is always connected to
the original standard output, and that's used exclusively for the
executed command output. The input is either UML 'null' channel or the
original stdin, depending on the --no-stdin flag.

The entire UML /dev/tty2 ('con2') may be a 'null' channel (discards
writes, blocks reads) if a mudem is not needed. If a mudem is needed,
it's hooked up to the pipes Python's subprocess.Popen creates for the
host side mudem process.

All other UML consoles are always connected to 'null' for input. The
output side is connected either to 'null' (by default) or the original
stderr (in verbose modes). This allows viewing the UML kernel and init
messages on the default console.

The UML kernel will always print some initial UML status messages,
either to stdout or stderr (depending on kernel version). Therefore
the subprocess stdout and stderr are redirected to /dev/null, or to
original stderr if in verbose mode. A dup of the original stdout is
passed to the UML process for use for the command output.