Externalize the umlbox timeout.
It turns out that time-keeping on UML is not an exact science: a wait
that's nominally supposed to be 30 seconds may take something like 10
minutes instead. Even if it was reliable, though, we would still need
a watchdog timer outside the UML system: if `init` breaks, or if the
child process manages to somehow mess with it, the command might
indefinitely block.

What this commit does is:

- It removes the previous `timeout` value communicated in the
  configuration proto into the running UML system, and the relevant
  code in init for dealing with the timers.

- It makes the umlbox wrapper script to instead take care of the
  timeout (when enabled). When the timeout expires, a soft timeout
  signal is sent to the UML by writing a '\x00' byte over the
  (otherwise so far unused) /dev/console input. If the UML process
  does not terminate in 5 seconds, a hard timeout signal is sent by
  writing a '\x01' byte. If it *still* does not terminate in 5
  seconds, it is forcibly stopped by delivering a KILL signal to the
  entire process group.

- It refactors the relevant run code in the umlbox init binary to
  watch for and to act on the soft timeout signals.