Fix timeout in case /console is not in raw mode.
Turns out without the `--verbose` flag, we never set /console to raw
mode, which causes the terminal driver (or whatnot) to apply line
buffering to it. This commit just does the cheap fix of changing the
timeout signal payloads from 0 and 1 to "N\n" and "Y\n", so that it
works in both modes. But maybe we should have used a separate device
altogether for the timeout signals.